+++ verify.go
@@ -6,7 +6,7 @@ import (
 )
 
 //returns true if the move is valid otherwise it returns false
-func chessVerify(source string, target string, promotion string, gameID int16) bool {
+func chessVerify(source string, target string, promotion string, gameID int) bool {
 	if len(source) != 2 {
 		fmt.Println("Invalid input length")
 		return false
@@ -223,7 +223,7 @@ func convertLetter(letter byte) int8 {
 }
 
 //makes the chess move on the board, verify the move first, returns captured piece as a string to be used in case of a move undo
-func makeMove(sourceRow int8, sourceCol int8, targetRow int8, targetCol int8, piece string, gameID int16) string {
+func makeMove(sourceRow int8, sourceCol int8, targetRow int8, targetCol int8, piece string, gameID int) string {
 
 	capturedPiece := Verify.AllTables[gameID].ChessBoard[targetRow][targetCol]
 	//make the source square blank as now the piece is no longer there
@@ -240,7 +240,7 @@ func makeMove(sourceRow int8, sourceCol int8, targetRow int8, targetCol int8, pi
 
 }
 
-func switchTurns(gameID int16) {
+func switchTurns(gameID int) {
 	if Verify.AllTables[gameID].whiteTurn == true {
 		Verify.AllTables[gameID].whiteTurn = false
 	} else {
@@ -248,7 +248,7 @@ func switchTurns(gameID int16) {
 	}
 }
 
-func printBoard(gameID int16) {
+func printBoard(gameID int) {
 
 	for i := 0; i < 8; i++ {
 		for j := 0; j < 8; j++ {
@@ -259,7 +259,7 @@ func printBoard(gameID int16) {
 }
 
 //undo a move if a player makes a move and doesn't stop check
-func undoMove(sourceRow int8, sourceCol int8, targetRow int8, targetCol int8, piece string, capturedPiece string, gameID int16) {
+func undoMove(sourceRow int8, sourceCol int8, targetRow int8, targetCol int8, piece string, capturedPiece string, gameID int) {
 	Verify.AllTables[gameID].ChessBoard[sourceRow][sourceCol] = piece
 	Verify.AllTables[gameID].ChessBoard[targetRow][targetCol] = capturedPiece
